+++ src/containers/App/App.css
@@ -1,7 +1,6 @@
 .App {
   font-family: 'Roboto Slab', serif;
-  height: -webkit-fill-available;
-}
+  height: -webkit-fill-available; }
 
 .toggle-map-style,
 .toggle-map-style-active {
@@ -10,12 +9,10 @@
   border-radius: 1rem;
   margin: auto;
   border: 2px solid #afafaf;
-  text-decoration-color: none;
-}
+  text-decoration-color: none; }
 
 button:focus {
-  outline: 0;
-}
+  outline: 0; }
 
 .toggle-map-style,
 .toggle-map-style-active {
@@ -24,42 +21,34 @@ button:focus {
   z-index: 10;
   right: 0;
   margin: 2.1rem 7rem 0 0;
-  top: 0;
-}
+  top: 0; }
 
 .toggle-map-style-active {
-  background-color: #5d5a5a;
-}
+  background-color: #5d5a5a; }
 
 .streets-button:hover {
-  background-color: gray;
-}
+  background-color: gray; }
 
 .streets-button,
 .dark-button {
   height: 1.55rem;
   width: 1.6rem;
   border-radius: 50%;
-  cursor: pointer;
-}
+  cursor: pointer; }
 
 .dark-button {
   background-color: gray;
   float: right;
-  cursor: pointer;
-}
+  cursor: pointer; }
 
 .dark-button:hover {
-  background-color: whitesmoke;
-}
+  background-color: whitesmoke; }
 
 .main-container {
-  display: flex;
-}
+  display: flex; }
 
 .dark {
-  color: white;
-}
+  color: white; }
 
 .quarter-circle-top-right {
   width: 60px;
@@ -74,149 +63,104 @@ button:focus {
   justify-content: center;
   align-items: center;
   border: 2px solid #717792;
-  transform: translate3d(-1.5rem, 1rem, 0);
-}
+  transform: translate3d(-1.5rem, 1rem, 0); }
 
 @font-face {
   font-family: codec;
-  src: url('../../fonts/codec.ttf');
-}
+  src: url("../../fonts/codec.ttf"); }
 
 @font-face {
   font-family: xtreme;
-  src: url('../../fonts/xtreem.ttf');
-}
+  src: url("../../fonts/xtreem.ttf"); }
 
 @media only screen and (max-width: 1000px) {
   .fa-stream {
-    font-size: 1.5rem;
-  }
+    font-size: 1.5rem; }
   .quarter-circle-top-right {
     height: 3rem;
-    width: 3rem;
-  }
-}
+    width: 3rem; } }
 
 @keyframes light-pulse {
   0% {
-    box-shadow: 0 0 0 0 rgba(228, 228, 228, 0.226);
-  }
+    box-shadow: 0 0 0 0 rgba(228, 228, 228, 0.226); }
   50% {
-    box-shadow: 0 0 0 7px rgba(207, 206, 206, 0.151);
-  }
+    box-shadow: 0 0 0 7px rgba(207, 206, 206, 0.151); }
   100% {
-    box-shadow: 0 0 0 0 rgba(251, 255, 255, 0.425);
-  }
-}
+    box-shadow: 0 0 0 0 rgba(251, 255, 255, 0.425); } }
 
 @keyframes dark-pulse {
   0% {
-    box-shadow: 0 0 0 0 rgba(252, 250, 250, 0);
-  }
+    box-shadow: 0 0 0 0 rgba(252, 250, 250, 0); }
   50% {
-    box-shadow: 0 0 0 20px rgba(15, 15, 15, 0.247);
-  }
+    box-shadow: 0 0 0 20px rgba(15, 15, 15, 0.247); }
   100% {
-    box-shadow: 0 0 0 0 rgba(251, 255, 255, 0.082);
-  }
-}
+    box-shadow: 0 0 0 0 rgba(251, 255, 255, 0.082); } }
 
 @keyframes fadeIn {
   from {
-    opacity: 0;
-  }
+    opacity: 0; }
   to {
-    opacity: 1;
-  }
-}
+    opacity: 1; } }
 
 @keyframes slideInRight {
   from {
     -webkit-transform: translate3d(100%, 0, 0);
     transform: translate3d(100%, 0, 0);
-    visibility: visible;
-  }
+    visibility: visible; }
   to {
     -webkit-transform: translate3d(0, 0, 0);
-    transform: translate3d(0, 0, 0);
-  }
-}
+    transform: translate3d(0, 0, 0); } }
 
 @keyframes slideInLeft {
   from {
     -webkit-transform: translate3d(-100%, 0, 0);
     transform: translate3d(-100%, 0, 0);
-    visibility: visible;
-  }
+    visibility: visible; }
   to {
     -webkit-transform: translate3d(0, 0, 0);
-    transform: translate3d(0, 0, 0);
-  }
-}
+    transform: translate3d(0, 0, 0); } }
 
 @keyframes slideInDown {
   from {
     -webkit-transform: translate3d(0, -100%, 0);
     transform: translate3d(0, -100%, 0);
-    visibility: visible;
-  }
+    visibility: visible; }
   to {
     -webkit-transform: translate3d(0, 0, 0);
-    transform: translate3d(0, 0, 0);
-  }
-}
+    transform: translate3d(0, 0, 0); } }
 
 @keyframes slideOutUp {
   from {
     -webkit-transform: translate3d(0, 0, 0);
-    transform: translate3d(0, 0, 0);
-  }
+    transform: translate3d(0, 0, 0); }
   to {
     visibility: hidden;
     -webkit-transform: translate3d(0, -100%, 0);
-    transform: translate3d(0, -100%, 0);
-  }
-}
+    transform: translate3d(0, -100%, 0); } }
 
 @keyframes flip {
   from {
-    -webkit-transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0)
-      rotate3d(0, 1, 0, -360deg);
-    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0)
-      rotate3d(0, 1, 0, -360deg);
+    -webkit-transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0) rotate3d(0, 1, 0, -360deg);
+    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0) rotate3d(0, 1, 0, -360deg);
     -webkit-animation-timing-function: ease-out;
-    animation-timing-function: ease-out;
-  }
+    animation-timing-function: ease-out; }
   40% {
-    -webkit-transform: perspective(400px) scale3d(1, 1, 1)
-      translate3d(0, 0, 150px) rotate3d(0, 1, 0, -190deg);
-    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 150px)
-      rotate3d(0, 1, 0, -190deg);
+    -webkit-transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 150px) rotate3d(0, 1, 0, -190deg);
+    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 150px) rotate3d(0, 1, 0, -190deg);
     -webkit-animation-timing-function: ease-out;
-    animation-timing-function: ease-out;
-  }
+    animation-timing-function: ease-out; }
   50% {
-    -webkit-transform: perspective(400px) scale3d(1, 1, 1)
-      translate3d(0, 0, 150px) rotate3d(0, 1, 0, -170deg);
-    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 150px)
-      rotate3d(0, 1, 0, -170deg);
+    -webkit-transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 150px) rotate3d(0, 1, 0, -170deg);
+    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 150px) rotate3d(0, 1, 0, -170deg);
     -webkit-animation-timing-function: ease-in;
-    animation-timing-function: ease-in;
-  }
+    animation-timing-function: ease-in; }
   80% {
-    -webkit-transform: perspective(400px) scale3d(0.95, 0.95, 0.95)
-      translate3d(0, 0, 0) rotate3d(0, 1, 0, 0deg);
-    transform: perspective(400px) scale3d(0.95, 0.95, 0.95) translate3d(0, 0, 0)
-      rotate3d(0, 1, 0, 0deg);
+    -webkit-transform: perspective(400px) scale3d(0.95, 0.95, 0.95) translate3d(0, 0, 0) rotate3d(0, 1, 0, 0deg);
+    transform: perspective(400px) scale3d(0.95, 0.95, 0.95) translate3d(0, 0, 0) rotate3d(0, 1, 0, 0deg);
     -webkit-animation-timing-function: ease-in;
-    animation-timing-function: ease-in;
-  }
+    animation-timing-function: ease-in; }
   to {
-    -webkit-transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0)
-      rotate3d(0, 1, 0, 0deg);
-    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0)
-      rotate3d(0, 1, 0, 0deg);
+    -webkit-transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0) rotate3d(0, 1, 0, 0deg);
+    transform: perspective(400px) scale3d(1, 1, 1) translate3d(0, 0, 0) rotate3d(0, 1, 0, 0deg);
     -webkit-animation-timing-function: ease-in;
-    animation-timing-function: ease-in;
-  }
-}
+    animation-timing-function: ease-in; } }
